Separate concurrent-safe code from non-concurrent-safe code
